refactor(coding): remove bash controller adapter
